modularize helper functions out of zipFiles
