+++ src/client/helperFunctions/capitalizeFirstLetter.ts
@@ -1,0 +1,4 @@
+// helper function that returns a string input with its first letter capitalized
+export function capitalizeFirstLetter(tag: string) {
+  return tag.charAt(0).toUpperCase() + tag.slice(1);
+}
